Refactor Pilot to isolate authn dependencies (havent done other modules yet)
