+++ pilot/pkg/authn/authentication_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package model
+package authn
 
 import (
 	"reflect"
