+++ pilot/pkg/security/authn/factory/factory.go
@@ -15,6 +15,7 @@
 package factory
 
 import (
+	authn2 "istio.io/istio/pilot/pkg/authn"
 	"istio.io/istio/pilot/pkg/model"
 	"istio.io/istio/pilot/pkg/security/authn"
 	"istio.io/istio/pilot/pkg/security/authn/v1alpha1"
@@ -25,6 +26,6 @@ import (
 func NewPolicyApplier(configStore model.IstioConfigStore,
 	serviceInstance *model.ServiceInstance) authn.PolicyApplier {
 	// TODO: check v1alpha2 policy and returns alpha2 applier, if exists.
-	authnPolicy := model.GetConsolidateAuthenticationPolicy(configStore, serviceInstance)
+	authnPolicy := authn2.GetConsolidateAuthenticationPolicy(configStore, serviceInstance)
 	return v1alpha1.NewPolicyApplier(authnPolicy)
 }
